Made a program that prints the numbers above main diagonal of a matrix
And here I am again, after a few days studying memory allocation and other low level things.
I decided to keep praticing the basics of C and made all the chalanges that I leave behind in the course (just because I was making the game for the final course project).